Move `patch_bitsandbytes_cuda` into `Engine`
